Payment_Service/requirements.txt
distlib==0.3.9
filelock==3.16.1
phonenumbers==8.13.51
platformdirs==4.3.6
virtualenv==20.28.0

# Django and Django REST Framework
django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.1
django-environ==0.10.0
# CORS Handling
django-cors-headers==4.6.0
# JWT Authentication
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0

# Database & ORM Support
psycopg2-binary==2.9.9     # If using MySQL

# Utility Libraries
python-dotenv==1.0.1  # For managing environment variables
pillow==10.2.0
twilio==8.9.0
requests==2.31.0
pika==1.3.2
celery
puka
razorpay